20200320:修复在Linux平台下编译问题，优化日志打印时间显示，版本号更新至0.0.3

+++ base/misc.h
@@ -22,7 +22,7 @@ typedef struct tag_ListEntry {
         (_list) = (_list)->next, (_cur) = (_list))
 
 #define MapTheListEntry(type, entry, member) \
-    ((type*)((entry) - (int)(&((type*)(0))->member)))
+    ((type*)((entry) - (unsigned long long)(&((type*)(0))->member)))
 
 void INTF_MISC_FreeDobulePoint(char **point, int num);
 int INTF_MISC_DivideStringBySymbol(char *srcStr, char symbol, char ***destStr, int *strNum);
